Split out the dynamodb client from chunk_store.go
