+++ chunk/chunk_store.go
@@ -3,15 +3,12 @@ package chunk
 import (
 	"encoding/json"
 	"fmt"
-	"math/rand"
 	"sort"
 	"strconv"
-	"sync"
 	"sync/atomic"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/prometheus/client_golang/prometheus"
@@ -30,52 +27,11 @@ const (
 	rangeKey = "r"
 	chunkKey = "c"
 
-	// For dynamodb errors
-	errorReasonLabel = "error"
-	otherError       = "other"
-
-	// Backoff for dynamoDB requests, to match AWS lib - see:
-	// https://github.com/aws/aws-sdk-go/blob/master/service/dynamodb/customizations.go
-	minBackoff = 50 * time.Millisecond
-	maxBackoff = 50 * time.Second
-
-	// Number of synchronous dynamodb requests
-	numDynamoRequests = 25
-
-	// See http://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Limits.html.
-	dynamoMaxBatchSize = 25
-
-	provisionedThroughputExceededException = "ProvisionedThroughputExceededException"
-
 	secondsInHour = int64(time.Hour / time.Second)
 	secondsInDay  = int64(24 * time.Hour / time.Second)
 )
 
 var (
-	dynamoRequestDuration = prometheus.NewHistogramVec(prometheus.HistogramOpts{
-		Namespace: "cortex",
-		Name:      "dynamo_request_duration_seconds",
-		Help:      "Time spent doing DynamoDB requests.",
-
-		// DynamoDB latency seems to range from a few ms to a few sec and is
-		// important.  So use 8 buckets from 64us to 8s.
-		Buckets: prometheus.ExponentialBuckets(0.000128, 4, 8),
-	}, []string{"operation", "status_code"})
-	dynamoConsumedCapacity = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: "cortex",
-		Name:      "dynamo_consumed_capacity_total",
-		Help:      "The capacity units consumed by operation.",
-	}, []string{"operation"})
-	dynamoFailures = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: "cortex",
-		Name:      "dynamo_failures_total",
-		Help:      "The total number of errors while storing chunks to the chunk store.",
-	}, []string{errorReasonLabel})
-	dynamoUnprocessedItems = prometheus.NewCounter(prometheus.CounterOpts{
-		Namespace: "cortex",
-		Name:      "dynamo_unprocessed_items_total",
-		Help:      "Unprocessed items",
-	})
 	indexEntriesPerChunk = prometheus.NewHistogram(prometheus.HistogramOpts{
 		Namespace: "cortex",
 		Name:      "chunk_store_index_entries_per_chunk",
@@ -116,13 +72,8 @@ var (
 )
 
 func init() {
-	prometheus.MustRegister(dynamoRequestDuration)
-	prometheus.MustRegister(dynamoConsumedCapacity)
-	prometheus.MustRegister(dynamoFailures)
-	prometheus.MustRegister(dynamoUnprocessedItems)
 	prometheus.MustRegister(indexEntriesPerChunk)
 	prometheus.MustRegister(s3RequestDuration)
-
 	prometheus.MustRegister(queryChunks)
 	prometheus.MustRegister(queryDynamoLookups)
 	prometheus.MustRegister(queryRequestPages)
@@ -165,29 +116,20 @@ type PeriodicTableConfig struct {
 type AWSStore struct {
 	cfg StoreConfig
 
-	dynamoRequests     chan dynamoOp
-	dynamoRequestsDone sync.WaitGroup
+	dynamo *dynamoDBBackoffClient
 }
 
 // NewAWSStore makes a new ChunkStore
-func NewAWSStore(cfg StoreConfig) (*AWSStore, error) {
-	store := &AWSStore{
-		cfg:            cfg,
-		dynamoRequests: make(chan dynamoOp),
+func NewAWSStore(cfg StoreConfig) *AWSStore {
+	return &AWSStore{
+		cfg:    cfg,
+		dynamo: newDynamoDBBackoffClient(cfg.DynamoDB),
 	}
-
-	store.dynamoRequestsDone.Add(numDynamoRequests)
-	for i := 0; i < numDynamoRequests; i++ {
-		go store.dynamoRequestLoop()
-	}
-
-	return store, nil
 }
 
 // Stop background goroutines.
 func (c *AWSStore) Stop() {
-	close(c.dynamoRequests)
-	c.dynamoRequestsDone.Wait()
+	c.dynamo.Stop()
 }
 
 type bucketSpec struct {
@@ -265,14 +207,6 @@ func parseRangeValue(v []byte) (label model.LabelName, value model.LabelValue, c
 	return
 }
 
-func recordDynamoError(err error) {
-	if awsErr, ok := err.(awserr.Error); ok {
-		dynamoFailures.WithLabelValues(awsErr.Code()).Add(float64(1))
-	} else {
-		dynamoFailures.WithLabelValues(otherError).Add(float64(1))
-	}
-}
-
 // Put implements ChunkStore
 func (c *AWSStore) Put(ctx context.Context, chunks []Chunk) error {
 	userID, err := user.GetID(ctx)
@@ -341,7 +275,7 @@ func (c *AWSStore) updateIndex(ctx context.Context, userID string, chunks []Chun
 		return err
 	}
 
-	return c.batchWriteDynamo(ctx, writeReqs)
+	return c.dynamo.batchWriteDynamo(ctx, writeReqs)
 }
 
 // calculateDynamoWrites creates a set of batched WriteRequests to dynamo for all
@@ -551,7 +485,7 @@ func (c *AWSStore) lookupChunksForMetricName(ctx context.Context, userID string,
 		queryDroppedMatches.Observe(float64(totalDropped))
 	}()
 
-	if err := c.queryPages(ctx, input, func(resp interface{}, lastPage bool) (shouldContinue bool) {
+	if err := c.dynamo.queryPages(ctx, input, func(resp interface{}, lastPage bool) (shouldContinue bool) {
 		var dropped int
 		dropped, processingError = processResponse(resp.(*dynamodb.QueryOutput), &chunkSet, nil)
 		totalDropped += dropped
@@ -609,7 +543,7 @@ func (c *AWSStore) lookupChunksForMatcher(ctx context.Context, userID string, bu
 		queryRequestPages.Observe(float64(pages))
 		queryDroppedMatches.Observe(float64(totalDropped))
 	}()
-	if err := c.queryPages(ctx, input, func(resp interface{}, lastPage bool) (shouldContinue bool) {
+	if err := c.dynamo.queryPages(ctx, input, func(resp interface{}, lastPage bool) (shouldContinue bool) {
 		var dropped int
 		dropped, processingError = processResponse(resp.(*dynamodb.QueryOutput), &chunkSet, matcher)
 		totalDropped += dropped
@@ -705,197 +639,4 @@ func (c *AWSStore) fetchChunkData(ctx context.Context, userID string, chunkSet [
 		return nil, errors[0]
 	}
 	return chunks, nil
-}
-
-// batchWriteDynamo writes many requests to dynamo in a single batch.
-func (c *AWSStore) batchWriteDynamo(ctx context.Context, reqs map[string][]*dynamodb.WriteRequest) error {
-	req := &dynamoBatchWriteItemsOp{
-		ctx:      ctx,
-		reqs:     reqs,
-		dynamodb: c.cfg.DynamoDB,
-		done:     make(chan error),
-	}
-	c.dynamoRequests <- req
-	return <-req.done
-}
-
-func (c *AWSStore) queryPages(ctx context.Context, input *dynamodb.QueryInput, callback func(resp interface{}, lastPage bool) (shouldContinue bool)) error {
-	page, _ := c.cfg.DynamoDB.QueryRequest(input)
-	req := &dynamoQueryPagesOp{
-		ctx:      ctx,
-		request:  page,
-		callback: callback,
-		done:     make(chan error),
-	}
-	c.dynamoRequests <- req
-	return <-req.done
-}
-
-func (c *AWSStore) dynamoRequestLoop() {
-	defer c.dynamoRequestsDone.Done()
-	for {
-		select {
-		case request, ok := <-c.dynamoRequests:
-			if !ok {
-				return
-			}
-			request.do()
-		}
-	}
-}
-
-type dynamoOp interface {
-	do()
-}
-
-type dynamoQueryPagesOp struct {
-	ctx      context.Context
-	request  dynamoRequest
-	callback func(resp interface{}, lastPage bool) (shouldContinue bool)
-	done     chan error
-}
-
-type dynamoBatchWriteItemsOp struct {
-	ctx      context.Context
-	reqs     map[string][]*dynamodb.WriteRequest
-	dynamodb DynamoDBClient
-	done     chan error
-}
-
-func nextBackoff(lastBackoff time.Duration) time.Duration {
-	// Based on the "Decorrelated Jitter" approach from https://www.awsarchitectureblog.com/2015/03/backoff.html
-	// sleep = min(cap, random_between(base, sleep * 3))
-	backoff := minBackoff + time.Duration(rand.Int63n(int64((lastBackoff*3)-minBackoff)))
-	if backoff > maxBackoff {
-		backoff = maxBackoff
-	}
-	return backoff
-}
-
-func (r *dynamoQueryPagesOp) do() {
-	backoff := minBackoff
-
-	for page := r.request; page != nil; page = page.NextPage() {
-		err := instrument.TimeRequestHistogram(r.ctx, "DynamoDB.QueryPages", dynamoRequestDuration, func(_ context.Context) error {
-			return page.Send()
-		})
-
-		if cc := page.Data().(*dynamodb.QueryOutput).ConsumedCapacity; cc != nil {
-			dynamoConsumedCapacity.WithLabelValues("DynamoDB.QueryPages").
-				Add(float64(*cc.CapacityUnits))
-		}
-
-		if err != nil {
-			recordDynamoError(err)
-
-			if awsErr, ok := err.(awserr.Error); ok && awsErr.Code() == provisionedThroughputExceededException {
-				time.Sleep(backoff)
-				backoff = nextBackoff(backoff)
-				continue
-			}
-
-			r.done <- page.Error()
-			return
-		}
-
-		if getNextPage := r.callback(page.Data(), !page.HasNextPage()); !getNextPage {
-			r.done <- page.Error()
-			return
-		}
-
-		backoff = minBackoff
-	}
-
-	r.done <- nil
-}
-
-func (r *dynamoBatchWriteItemsOp) do() {
-	min := func(i, j int) int {
-		if i < j {
-			return i
-		}
-		return j
-	}
-
-	dictLen := func(in map[string][]*dynamodb.WriteRequest) int {
-		result := 0
-		for _, reqs := range in {
-			result += len(reqs)
-		}
-		return result
-	}
-
-	// Fill 'out' with WriteRequests from 'in' until it 'out' has at most dynamoMaxBatchSize requests. Remove those requests from 'in'.
-	fillReq := func(in map[string][]*dynamodb.WriteRequest, out map[string][]*dynamodb.WriteRequest) {
-		outLen, inLen := dictLen(out), dictLen(in)
-		toFill := min(inLen, dynamoMaxBatchSize-outLen)
-		for toFill > 0 {
-			for tableName := range in {
-				reqs := in[tableName]
-				taken := min(len(reqs), toFill)
-				out[tableName] = append(out[tableName], reqs[:taken]...)
-				in[tableName] = reqs[taken:]
-				toFill -= taken
-			}
-		}
-	}
-
-	copyUnprocessed := func(in map[string][]*dynamodb.WriteRequest, out map[string][]*dynamodb.WriteRequest) {
-		for tableName, unprocessReqs := range in {
-			out[tableName] = append(out[tableName], unprocessReqs...)
-			dynamoUnprocessedItems.Add(float64(len(unprocessReqs)))
-		}
-	}
-
-	outstanding, unprocessed := r.reqs, map[string][]*dynamodb.WriteRequest{}
-	backoff := minBackoff
-	for dictLen(outstanding)+dictLen(unprocessed) > 0 {
-		reqs := map[string][]*dynamodb.WriteRequest{}
-		fillReq(unprocessed, reqs)
-		fillReq(outstanding, reqs)
-
-		var resp *dynamodb.BatchWriteItemOutput
-		err := instrument.TimeRequestHistogram(r.ctx, "DynamoDB.BatchWriteItem", dynamoRequestDuration, func(_ context.Context) error {
-			var err error
-			resp, err = r.dynamodb.BatchWriteItem(&dynamodb.BatchWriteItemInput{
-				RequestItems:           reqs,
-				ReturnConsumedCapacity: aws.String(dynamodb.ReturnConsumedCapacityTotal),
-			})
-			if err != nil {
-				recordDynamoError(err)
-			}
-			return err
-		})
-		for _, cc := range resp.ConsumedCapacity {
-			dynamoConsumedCapacity.WithLabelValues("DynamoDB.BatchWriteItem").
-				Add(float64(*cc.CapacityUnits))
-		}
-
-		// If there are unprocessed items, backoff and retry those items.
-		if resp.UnprocessedItems != nil && dictLen(resp.UnprocessedItems) > 0 {
-			copyUnprocessed(resp.UnprocessedItems, unprocessed)
-			time.Sleep(backoff)
-			backoff = nextBackoff(backoff)
-			continue
-		}
-
-		// If we get provisionedThroughputExceededException, then no items were processed,
-		// so back off and retry all.
-		if awsErr, ok := err.(awserr.Error); ok && awsErr.Code() == provisionedThroughputExceededException {
-			copyUnprocessed(reqs, unprocessed)
-			time.Sleep(backoff)
-			backoff = nextBackoff(backoff)
-			continue
-		}
-
-		// All other errors are fatal.
-		if err != nil {
-			r.done <- err
-			return
-		}
-
-		backoff = minBackoff
-	}
-
-	r.done <- nil
 }
